Updating windows and linux scripts

+++ scripts/linux/make_ifort_pycrysfml08.sh
@@ -31,26 +31,75 @@
 #
 # -------------------------------------------------------------
 
-echo Building CFML_Sxtal_Geom
-ifort -fPIC -fpp -c ../../src/py_cfml_sxtal_geom.f90 -I$CRYSFML08_INSTALL/libC
-ifort -shared -o py_cfml_sxtal_geom.so py_cfml_sxtal_geom.o -L $CRYSFML08_INSTALL/libC -l crysfml08
-mv py_cfml_sxtal_geom.so ../../pycrysfml08/
+INSTALLATION_DIR=""
+CRYSFML08_INCLUDE_DIR=""
+CRYSFML08_LIB_DIR=""
 
-echo Building CFML_DiffPatt
-ifort -fPIC -fpp -c ../../src/py_extension_cfml_diffpatt.f90 -I$CRYSFML08_INSTALL/libC
-ifort -fPIC -fpp -c ../../src/py_cfml_diffpatt.f90 -I$CRYSFML08_INSTALL/libC
-ifort -shared -o py_cfml_diffpatt.so py_extension_cfml_diffpatt.o py_cfml_diffpatt.o -L $CRYSFML08_INSTALL/libC -l crysfml08
-mv py_cfml_diffpatt.so ../../pycrysfml08/
+compile=true
+if [ -z "$INSTALLATION_DIR"]
+then
+    echo "Please set INSTALLATION_DIR variable in the script"
+    compile=false
+fi
+if [ -z "$CRYSFML08_INCLUDE_DIR"]
+then
+    echo "Please set CRYSFML08_INCLUDE_DIR variable in the script"
+    compile=false
+fi
+if [ -z "$CRYSFML08_LIB_DIR"]
+then
+    echo "Please set CRYSFML08_LIB_DIR variable in the script"
+    compile=false
+fi
 
-echo Building CFML_Reflections
-ifort -fPIC -fpp -c ../../src/py_extension_cfml_reflections.f90 -I$CRYSFML08_INSTALL/libC
-ifort -fPIC -fpp -c ../../src/py_cfml_reflections.f90 -I$CRYSFML08_INSTALL/libC
-ifort -shared -o py_cfml_reflections.so py_extension_cfml_reflections.o py_cfml_reflections.o -L $CRYSFML08_INSTALL/libC -l crysfml08
-mv py_cfml_reflections.so ../../pycrysfml08/
+if $compile
+then
+    INSTALLATION_DIR=$INSTALLATION_DIR/pycrysfml08
 
-echo Building CFML_VTK
-ifort -fPIC -fpp -c ../../src/py_cfml_vtk.f90 -I$CRYSFML08_INSTALL/libC
-ifort -shared -o py_cfml_vtk.so py_cfml_vtk.o -L $CRYSFML08_INSTALL/libC -l crysfml08
-mv py_cfml_vtk.so ../../pycrysfml08/
+    echo Building CFML_Sxtal_Geom
+    ifort -fPIC -fpp -c ../../src/py_cfml_sxtal_geom.f90 -I$CRYSFML08_INCLUDE_DIR
+    ifort -shared -o py_cfml_sxtal_geom.so py_cfml_sxtal_geom.o -L $CRYSFML08_LIB_DIR -l crysfml08
+    mv py_cfml_sxtal_geom.so $INSTALLATION_DIR
 
-rm *.o *.mod
+    echo Building CFML_DiffPatt
+    ifort -fPIC -fpp -c ../../src/py_cfml_diffpatt.f90 -I$CRYSFML08_INCLUDE_DIR
+    ifort -shared -o py_cfml_diffpatt.so py_cfml_diffpatt.o -L $CRYSFML08_LIB_DIR -l crysfml08
+    mv py_cfml_diffpatt.so $INSTALLATION_DIR
+
+    echo Building CFML_gSpaceGroups
+    ifort -fPIC -fpp -c ../../src/py_cfml_spacegroups.f90 -I$CRYSFML08_INCLUDE_DIR
+    ifort -shared -o py_cfml_spacegroups.so py_cfml_spacegroups.o -L $CRYSFML08_LIB_DIR -l crysfml08
+    mv py_cfml_spacegroups.so $INSTALLATION_DIR
+
+    echo Building CFML_IOForm
+    ifort -fPIC -fpp -c ../../src/py_cfml_ioform.f90 -I$CRYSFML08_INCLUDE_DIR
+    ifort -shared -o py_cfml_ioform.so py_cfml_ioform.o -L $CRYSFML08_LIB_DIR -l crysfml08
+    mv py_cfml_ioform.so $INSTALLATION_DIR
+
+    echo Building CFML_Metrics
+    ifort -fPIC -fpp -c ../../src/py_cfml_metrics.f90 -I$CRYSFML08_INCLUDE_DIR
+    ifort -shared -o py_cfml_metrics.so py_cfml_metrics.o -L $CRYSFML08_LIB_DIR -l crysfml08
+    mv py_cfml_metrics.so $INSTALLATION_DIR
+
+    echo Building CFML_Profiles
+    ifort -fPIC -fpp -c ../../src/py_cfml_profiles.f90 -I$CRYSFML08_INCLUDE_DIR
+    ifort -shared -o py_cfml_profiles.so py_cfml_profiles.o -L $CRYSFML08_LIB_DIR -l crysfml08
+    mv py_cfml_profiles.so $INSTALLATION_DIR
+
+    echo Building CFML_Reflections
+    ifort -fPIC -fpp -c ../../src/py_cfml_reflections.f90 -I$CRYSFML08_INCLUDE_DIR
+    ifort -shared -o py_cfml_reflections.so py_cfml_reflections.o -L $CRYSFML08_LIB_DIR -l crysfml08
+    mv py_cfml_reflections.so $INSTALLATION_DIR
+
+    echo Building CFML_Structure_Factors
+    ifort -fPIC -fpp -c ../../src/py_cfml_structure_factors.f90 -I$CRYSFML08_INCLUDE_DIR
+    ifort -shared -o py_cfml_structure_factors.so py_cfml_structure_factors.o -L $CRYSFML08_LIB_DIR -l crysfml08
+    mv py_cfml_structure_factors.so $INSTALLATION_DIR
+
+    echo Building CFML_Sxtal_Geom
+    ifort -fPIC -fpp -c ../../src/py_cfml_sxtal_geom.f90 -I$CRYSFML08_INCLUDE_DIR
+    ifort -shared -o py_cfml_sxtal_geom.so py_cfml_sxtal_geom.o -L $CRYSFML08_LIB_DIR -l crysfml08
+    mv py_cfml_sxtal_geom.so $INSTALLATION_DIR
+
+    rm *.o *.mod
+fi
